corrective-rag/pyproject.toml
[project]
name = "corrective-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "black>=26.1.0",
    "chromadb>=1.5.7",
    "isort>=8.0.1",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.10",
    "langchainhub>=0.1.21",
    "langgraph>=1.0.10",
    "pytest>=9.0.3",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.23",
]
